Move dml handler under context and use absolute paths
