+++ runner/jest.config.js
@@ -1,5 +1,6 @@
 module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
-  roots: ['./src', './tests'],
+  roots: ['./'],
+  modulePaths: ['<rootDir>'],
 };
