--- runner/src/dml-handler/index.ts
@@ -1,1 +1,0 @@
-export { default } from './dml-handler';
