+++ runner/src/indexer/context-builder/context-builder.test.ts
@@ -1,7 +1,7 @@
 import type fetch from 'node-fetch';
 
-import type DmlHandler from '../../dml-handler';
-import IndexerConfig from '../../indexer-config';
+import type { DmlHandler } from './dml-handler';
+import IndexerConfig from 'src/indexer-config';
 import { LogLevel } from '../../indexer-meta/log-entry';
 import ContextBuilder from './context-builder';
