+++ runner/src/indexer/context-builder/dml-handler/in-memory-dml-handler.ts
@@ -1,5 +1,5 @@
 import { type AST, Parser } from 'node-sql-parser';
-import { type TableDefinitionNames } from '../indexer';
+import { type TableDefinitionNames } from 'src/indexer';
 import { type PostgresRow, type WhereClauseMulti, type WhereClauseSingle, type DmlHandlerInterface } from './dml-handler';
 
 // TODO: Define class to represent specification
