+++ runner/tsconfig.json
@@ -3,6 +3,7 @@
     "target": "es2018",                                  /* Set the JavaScript language version for emitted JavaScript and include compatible library declarations. */
     "lib": ["es2021"],                                      /* Specify a set of bundled library declaration files that describe the target runtime environment. */
     "module": "commonjs",                                /* Specify what module code is generated. */
+    "baseUrl": "./",                                     /* Base directory to resolve non-absolute module names. */
     "rootDirs": ["./src", "./tests", "./scripts"],    
     "moduleResolution": "node",                     /* Specify how TypeScript looks up a file from a given module specifier. */
     "resolveJsonModule": true,                           /* Enable importing .json files. */
